+++ crates/core/src/json/client_api.rs
@@ -81,7 +81,7 @@ pub struct EventJson {
     pub status: String,          // committed, failed
     pub caller_identity: String, // hex identity
     pub function_call: FunctionCallJson,
-    pub energy_quanta_used: u64,
+    pub energy_quanta_used: i128,
     pub message: String,
 }
